hackney v1.15.2 hackney_bstr

Types

cp()
(opaque)

-opaque cp() :: {am | bm, binary()}.

part()

part() :: {Start :: non_neg_integer(), Length :: integer()}.

END: Remove when OTP 17 not officially supported

Functions

alpha/2

alpha(binary(), fun()) -> any().

Parse a list of case-insensitive alpha characters.

Changes all characters to lowercase.

char_to_lower/1

char_to_lower(char()) -> char().

Convert [A-Z] characters to lowercase.

char_to_upper/1

char_to_upper(char()) -> char().

Convert [a-z] characters to uppercase.

digits/1

digits(binary()) -> non_neg_integer() | {error, badarg}.

Parse a list of digits as a non negative integer.

digits/2

digits(binary(), fun()) -> any().

digits/3

digits(binary(), fun(), non_neg_integer()) -> any().

join/2

list/2

list(binary(), fun()) -> list() | {error, badarg}.

Parse a list of the given type.

nonempty_list/2

nonempty_list(binary(), fun()) -> [any(), ...] | {error, badarg}.

Parse a non-empty list of the given type.

parameterized_tokens/1

parameterized_tokens(binary()) -> any().

Parse a non empty list of tokens followed with optional parameters.

params/2

params(binary(), fun()) -> any().

Parse a list of parameters (a=b;c=d).

quoted_string/2

quoted_string(binary(), fun()) -> any().

split/3

split(Subject, Pattern, Options) -> Parts
         when
             Subject :: binary(),
             Pattern :: binary() | [binary()] | cp(),
             Options :: [Option],
             Option :: {scope, part()} | trim | global | trim_all,
             Parts :: [binary()].

to_binary/1

to_hex/1

to_lower/1

to_lower(binary() | atom() | list()) -> binary().

Convert a binary string to lowercase.

to_upper/1

to_upper(binary() | atom() | list()) -> binary().

token/2

token(binary(), fun()) -> any().

Parse a token.

token_ci/2

token_ci(binary(), fun()) -> any().

Parse a case-insensitive token.

Changes all characters to lowercase.

trim/1

trim(binary()) -> binary().

whitespace/2

whitespace(binary(), fun()) -> any().

Skip whitespace.

word/2

word(binary(), fun()) -> any().

Parse either a token or a quoted string.