hackney v1.15.2 hackney_connect

Functions

check_or_close/1

close/1

close the client

connect/3

connect/4

connect/5

create_connection/4

create a connection and return a client state

create_connection/5

is_pool/1

get current pool pid or name used by a client if needed

maybe_connect/1

connect a socket and create a client state.

partial_chain/1

peername/1

get the address and port for the other end of current connection in the client

reconnect/4

set_sockopts/2

add set sockets options in the client

sockname/1

the local address and port of current socket in the client

ssl_opts/2