hackney v1.15.2 hackney_headers

module to manipulate HTTP headers

Types

disposition()

disposition() :: {binary(), [{binary(), binary()}]}.

headers()

headers() :: any().

Functions

content_disposition/1

content_disposition(binary()) -> disposition().

Parse a content disposition.

content_type/1

content_type(binary()) -> any().

Parse a content type.

We lowercase the charset header as we know it's case insensitive.

delete/2

Delete the header corresponding to key if it is present.

fold/3

fold the list of headers

get_value/2

get the value of the header

get_value/3

header_value/2

join value and params in a binary

insert/3

Insert the pair into the headers, merging with any pre-existing key. A merge is done with Value = V0 ++ ", " ++ V1.

insert/4

same as insert/3 but allows to add params to the header value.

make_header/2

Create a binary header

make_header/3

new/0

new() -> headers().

initialise an header dict

new/1

new(list()) -> headers().

parse/2

parse(binary(), list() | headers()) ->
         any() | undefined | {error, badarg}.

Semantically parse headers.

When the value isn't found, a proper default value for the type returned is used as a return value.

store/3

store the pair into the headers, replacing any pre-existing key.

to_binary/1

return all the headers as a binary that can be sent over the wire.

to_list/1

update/2

extend the headers with a new list of {Key, Value} pair.