hackney v1.15.2 hackney_headers_new

Types

headers()

headers() :: term().

headers_list()

headers_list() :: [{key(), value()}].

key()

key() :: binary() | string().

value()

value() ::
    binary() | {binary() | [{binary(), binary()} | binary()]}.

Functions

append/3

append(key(), value(), headers()) -> headers().

append a new value to the list of value for the the header field if the key has not been recorded the list will be created with the value as the first item.

delete/2

delete(key(), headers()) -> headers().

delete a field from headers.

fold/3

from_list/1

from_list(headers_list()) -> headers().

create headers from a list

get_value/2

get_value(key(), headers()) -> value() | undefined.

get the first value of an headers or return undefined

get_value/3

get_value(key(), headers(), any()) -> value() | any().

get the first value of an headers or return the default

is_key/2

is_key(key(), headers()) -> true | false.

is the header field exists or no

lookup/2

merge/2

merge(headers(), headers()) -> headers().

merge 2 headers objects. If a key is already exising in HEader1, it will be kept.

new/0

new() -> headers().

initialize an empty headers objecy

new/1

new(headers_list() | headers()) -> headers().

parse_content_type/1

parse_content_type(binary()) -> any().

parse_media_type/2

parse_media_type(binary(), fun()) -> any().

Parse a media type.

size/1

size(headers()) -> non_neg_integer().

return the number of headers fields

store/2

store(headers_list(), headers()) -> headers().

store a list of headers. Replacing oldest

store/3

store(key(), value() | [value()], headers()) -> headers().

replace the content of the header field with the value or the list of values.

store_new/3

store_new(key(), value(), headers()) -> {boolean(), headers()}.

only store a vakue if the key exist.

to_binary/1

to_binary(headers()) -> binary().

transform headers to a binary that can be used to construct a request

to_iolist/1

to_iolist(headers()) -> iolist().

conver headers to an iolist. Useful to send them over the wire.

to_list/1

to_list(headers()) -> headers_list().

convert headers to a list