hackney v1.15.2 hackney_local_tcp

Functions

close/1

close(inet:socket()) -> ok.

Close a TCP socket.

connect/3

connect/4

controlling_process/2

controlling_process(inet:socket(), pid()) ->
                       ok | {error, closed | not_owner | atom()}.

Assign a new controlling process Pid to Socket.

messages/1

Atoms used to identify messages in {active, once | true} mode.

peername/1

peername(inet:socket()) ->
            {ok, {inet:ip_address(), inet:port_number()}} |
            {error, atom()}.

Return the address and port for the other end of a connection.

recv/2

recv/3

recv(inet:socket(), non_neg_integer(), timeout()) ->
        {ok, any()} | {error, closed | atom()}.

Receive a packet from a socket in passive mode.

send/2

send(inet:socket(), iolist()) -> ok | {error, atom()}.

Send a packet on a socket.

setopts/2

setopts(inet:socket(), list()) -> ok | {error, atom()}.

Set one or more options for a socket.

shutdown/2

shutdown(inet:socket(), read | write | read_write) -> ok.

Immediately close a socket in one or two directions.

sockname/1

sockname(inet:socket()) ->
            {ok, {inet:ip_address(), inet:port_number()}} |
            {error, atom()}.

Get the local address and port of a socket