hackney v1.15.2 hackney_pool

pool of sockets connections

Functions

checkin/2

release a socket in the pool

checkout/4

fetch a socket from the pool

child_spec/2

return a child spec suitable for embeding your pool in the supervisor

code_change/3

Callback implementation for c::gen_server.code_change/3.

count/1

get the number of connections in the pool

count/2

get the number of connections in the pool for {Host0, Port, Transport}

find_pool/1

get_stats/1

handle_call/3

Callback implementation for c::gen_server.handle_call/3.

handle_cast/2

Callback implementation for c::gen_server.handle_cast/2.

handle_info/2

Callback implementation for c::gen_server.handle_info/2.

init/1

Callback implementation for c::gen_server.init/1.

max_connections/1

get max pool size

notify/2

set_max_connections/2

change the pool size

set_timeout/2

change the connection timeout

start/0

start_link/2

start_pool/2

start a pool

stop_pool/1

stop a pool

terminate/2

Callback implementation for c::gen_server.terminate/2.

timeout/1

get timeout