hackney v1.15.2 hackney_url

module to manage URLs.

Types

hackney_url()

hackney_url() :: #hackney_url{}.

qs_opt()

qs_opt() :: noplus | upper.

qs_vals()

qs_vals() :: [{binary(), binary() | true}].

Functions

fix_path/1

idnconvert_hostname/1

make_url/3

make_url(binary(), binary() | [binary()], binary() | qs_vals()) ->
            binary().

Construct an URL from a base URL, a path and a list of properties to give to the URL.

normalize/1

normalize(URL) -> NormalizedUrl
             when
                 URL :: binary() | list() | hackney_url(),
                 NormalizedUrl :: hackney_url().

Normalizes the encoding of an URL. Use the hackney_url:pathencode/1 to encode an URL.

normalize/2

normalize(URL, Fun) -> NormalizedUrl
             when
                 URL :: binary() | list() | hackney_url(),
                 Fun :: fun(),
                 NormalizedUrl :: hackney_url().

Normalizes the encoding of an URL.

parse_qs/1

parse_qs(binary()) -> qs_vals().

parse_url/1

parse_url(URL :: binary() | list()) -> hackney_url().

Parse an URL and return a #hackney_url record.

pathencode/1

pathencode(binary()) -> binary().

qs/1

qs(qs_vals()) -> binary().

Encode query properties to binary.

qs/2

qs(qs_vals(), [qs_opt()]) -> binary().

Encode query properties to binary. Opts are passed to urlencode/2

transport_scheme/1

unparse_url/1

urldecode/1

urldecode(binary()) -> binary().

urldecode/2

urldecode(binary(), crash | skip) -> binary().

Decode an URL encoded binary. The second argument specifies how to handle percent characters that are not followed by two valid hex characters. Use skip to ignore such errors, if crash is used the function will fail with the reason badarg.

urlencode/1

urlencode(binary() | string()) -> binary().

URL encode a string binary.

urlencode/2

urlencode(binary() | string(), [qs_opt()]) -> binary().

URL encode a string binary. The noplus option disables the default behaviour of quoting space characters, \s, as +. The upper option overrides the default behaviour of writing hex numbers using lowecase letters to using uppercase letters instead.