hackney v1.15.2 hackney_util

Functions

filter_options/3

filter_options([{atom(), any()} | {raw, any(), any(), any()}],
               [atom()],
               Acc) ->
                  Acc
                  when Acc :: [any()].

filter a proplists and only keep allowed keys

is_ipv6/1

maybe_apply_defaults/2

merge_opts/2

mod_metrics/0

privdir/0

require/1

require([module()]) -> ok.

Start the given applications if they were not already started.

set_option_default/3

set_option_default(Opts, atom(), any()) -> Opts
                      when Opts :: [{atom(), any()}].

set the default options in a proplists if not defined

to_atom/1

to_int/1